Add Makefile for fully automated building
Building the project is fully automated. VDF and NSIS setup are created
in the sub-directory 'build' (excluded in .gitignore) by running 'make',
'make all', or 'make remake'. Any metadata like project name, version
number, website, etc. are updated automatically and are to be modified
in the file 'metadata.txt'. The project license will be added to the
NSIS setup from the file 'LICENSE'. This file is empty at time of this
commit.

To build, the paths to the following software are required to be in the
PATH environment variable:

 - Make
 - GothicVDFS
 - NSIS

The Makefile is written and supported for Windows only.