Add console commands to show short names of fixes
The console command `fix name xxx` with xxx being a fix ID shows the
short name of that fix. The command `fix name all` shows an ordered list
of all fixes with their short names and whether they are applied, not
applied or disabled.

The short name refers to the suffix of the function name that is
responsible for the fix.