refactor: add make release rule
The make rule "dev" triggers "vdf", while the rule "release" triggers
"vdf setup". Depending on the build type, "vdf" will omit the test
suite. The resulting files are built in the respective sub-directories
"/build/dev/" and "/build/release/". The default make target is "dev".
That means, running "make" from command line will create the complete
VDF (but not the setup) in the directory "/build/dev/". To create the
release (including the setup), run "make release" from command line.
Note that both dev and release build contain all source files. The only
difference is that the test suite will not be parsed in the release.