Extend finding VOBs further with callback
Instead of specifying a class to narrow down the search (see previous
commit), the option is now more versatile: Now the last parameter is a
callback function. It can either be one of the existing class check
functions or a custom function to narrow down the search more flexibly
and potentially based on more or more complex checks.

Refs #46 #51 #52 #127 #212 #213 #226