Rephrase and refactor NPC rename template
To clarify the use of the template, it is now differentiated between the
name of the NPC and the name of the NPC instance. Redundant comments in
the template are removed to avoid wall-of-text and emphasize the
simplicity to apply the template. In the "TODO" strings of the template,
the optional colon after "TODO" is removed, to reduce the characters on
the line.

Ref #415