Add template for rename NPC instance (#416)
* Add template for rename NPC instance

* Refactor comments in low-level string functions

* Implement unique string source identification

* Refactor constant assignments in tests

* Rephrase and refactor NPC rename template

To clarify the use of the template, it is now differentiated between the
name of the NPC and the name of the NPC instance. Redundant comments in
the template are removed to avoid wall-of-text and emphasize the
simplicity to apply the template. In the "TODO" strings of the template,
the optional colon after "TODO" is removed, to reduce the characters on
the line.

* Refactor #149 for unique replacement string

The test broke because of #145. Now fixed.

* Remove previously added empty line

* Fix broken tests from previous commits

Closes #415

---------

Co-authored-by: szapp <dev.szapp@gmail.com>