Add efficient array symbol name decomposition
Symbols holding data types with arrays require to equip all symbol
getter and setter functions with an additional parameter for the index.
That parameter is almost always zero (there are not many arrays), which
convolutes the long function signature and makes function calls
confusing.

The idea here is to pass the array index as part of the symbol name,
e.g. "count[4]" and to decompose this string into the symbol name
"count" and the array index 4.

This commit lays the ground work to simplify many function calls e.g.
from `G1CP_IsIntVar("Kapitel", 0)` to `G1CP_IsIntVar("Kapitel")` or
better from `G1CP_SetIntVar("InExtremoPlaying", 0, TRUE)` to a much less
obscure `G1CP_SetIntVar("InExtremoPlaying", TRUE)`. And for values in
arrays, it will be `G1CP_GetStringId("C_Item.text[1]")`.

The implementation is low-level to avoid any overhead on the string
operations.